1、修复cdn失效问题 2、同步博客最新功能

+++ src/main/java/com/star/controller/CommentController.java
@@ -1,6 +1,8 @@
 package com.star.controller;
 
+import com.star.annotation.AccessLimit;
 import com.star.entity.Comment;
+import com.star.entity.Message;
 import com.star.entity.User;
 import com.star.queryvo.DetailedBlog;
 import com.star.service.BlogService;
@@ -12,7 +14,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
@@ -20,8 +21,8 @@ import java.util.List;
 
 /**
  * @Description: 评论控制器
+ * @Date: Created in 10:25 2020/6/23
  * @Author: ONESTAR
- * @Date: Created in 17:16 2020/4/5
  * @QQ群: 530311074
  * @URL: https://onestar.newstar.net.cn/
  */
@@ -37,7 +38,7 @@ public class CommentController {
     @Value("${comment.avatar}")
     private String avatar;
 
-//    查询评论列表
+    //查询评论列表
     @GetMapping("/comments/{blogId}")
     public String comments(@PathVariable Long blogId, Model model) {
         List<Comment> comments = commentService.listCommentByBlogId(blogId);
@@ -45,9 +46,10 @@ public class CommentController {
         return "blog :: commentList";
     }
 
-//    新增评论
+    //新增评论
     @PostMapping("/comments")
-    public String post(Comment comment, HttpSession session,Model model) {
+    @AccessLimit(seconds = 15, maxCount = 3) //15秒内 允许请求3次
+    public String post(Comment comment, HttpSession session, Model model) {
         Long blogId = comment.getBlogId();
         User user = (User) session.getAttribute("user");
         if (user != null) {
@@ -57,20 +59,27 @@ public class CommentController {
             //设置头像
             comment.setAvatar(avatar);
         }
-
+        Long parentId = comment.getParentComment().getId();
+        Comment parentComment = null;
         if (comment.getParentComment().getId() != null) {
-            comment.setParentCommentId(comment.getParentComment().getId());
+            comment.setParentCommentId(parentId);
+
+            // 根据父评论id查询留言信息
+            parentComment = commentService.getEmailByParentId(parentId);
         }
-        commentService.saveComment(comment);
+        commentService.saveComment(comment,parentComment);
         List<Comment> comments = commentService.listCommentByBlogId(blogId);
         model.addAttribute("comments", comments);
         return "blog :: commentList";
     }
 
-//    删除评论
+    //删除评论
     @GetMapping("/comment/{blogId}/{id}/delete")
-    public String delete(@PathVariable Long blogId, @PathVariable Long id,Comment comment, RedirectAttributes attributes, Model model){
-        commentService.deleteComment(comment,id);
+    public String delete(@PathVariable Long blogId, @PathVariable Long id, Comment comment, HttpSession session, Model model){
+        User user = (User) session.getAttribute("user");
+        if(user != null) {
+            commentService.deleteComment(comment,id);
+        }
         DetailedBlog detailedBlog = blogService.getDetailedBlog(blogId);
         List<Comment> comments = commentService.listCommentByBlogId(blogId);
         model.addAttribute("blog", detailedBlog);
