+++ src/main/java/com/star/controller/MessageController.java
@@ -1,5 +1,8 @@
 package com.star.controller;
 
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
+import com.star.annotation.AccessLimit;
 import com.star.entity.Message;
 import com.star.entity.User;
 import com.star.service.MessageService;
@@ -7,10 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
@@ -33,21 +33,30 @@ public class MessageController {
     private String avatar;
 
     @GetMapping("/message")
-    public String message() {
+    public String message(Model model,@RequestParam(defaultValue = "1",value = "pageNum") Integer pageNum) {
+        PageHelper.startPage(pageNum,15);
+        List<Message> messages = messageService.listMessage();
+        PageInfo<Message> pageInfo = new PageInfo<Message>(messages);
+        model.addAttribute("messages", pageInfo);
         return "message";
     }
 
 //    查询留言
     @GetMapping("/messagecomment")
-    public String messages(Model model) {
+    public String messages(Model model,@RequestParam(defaultValue = "1",value = "pageNum") Integer pageNum) {
+
+        PageHelper.startPage(pageNum,15);
         List<Message> messages = messageService.listMessage();
-        model.addAttribute("messages", messages);
+        PageInfo<Message> pageInfo = new PageInfo<Message>(messages);
+        model.addAttribute("messages", pageInfo);
+
         return "message::messageList";
     }
 
 //    新增留言
     @PostMapping("/message")
-    public String post(Message message, HttpSession session, Model model) {
+    @AccessLimit(seconds = 15, maxCount = 3) //15秒内 允许请求3次
+    public String post(Message message, HttpSession session, Model model,@RequestParam(defaultValue = "1",value = "pageNum") Integer pageNum) {
         User user = (User) session.getAttribute("user");
         //设置头像
         if (user != null) {
@@ -56,20 +65,31 @@ public class MessageController {
         } else {
             message.setAvatar(avatar);
         }
-        if (message.getParentMessage().getId() != null) {
-            message.setParentMessageId(message.getParentMessage().getId());
+        Long parentId = message.getParentMessage().getId();
+        Message parentMessage = null;
+        if (parentId != null) {
+            message.setParentMessageId(parentId);
+            // 根据父评论id查询留言信息
+            parentMessage = messageService.getEmailByParentId(parentId);
         }
-        messageService.saveMessage(message);
+        messageService.saveMessage(message,parentMessage);
+
+        PageHelper.startPage(pageNum,15);
         List<Message> messages = messageService.listMessage();
-        model.addAttribute("messages", messages);
-        return "message::messageList";
+        PageInfo<Message> pageInfo = new PageInfo<Message>(messages);
+        model.addAttribute("messages", pageInfo);
+        return "message";
+
+        // return "redirect:/message";
     }
 
 //    删除留言
     @GetMapping("/messages/{id}/delete")
-    public String delete(@PathVariable Long id, RedirectAttributes attributes, Model model){
-        messageService.deleteMessage(id);
+    public String delete(@PathVariable Long id, RedirectAttributes attributes, Model model,HttpSession session){
+        User user = (User) session.getAttribute("user");
+        if(user != null){
+            messageService.deleteMessage(id);
+        }
         return "redirect:/message";
     }
-
 }
